+++ copy_with_extension_gen/test/gen_basic_functionality_test.dart
@@ -67,6 +67,11 @@ void main() {
       const CopyWithValuesOptional(id: "test").copyWith.id(null).id,
       null,
     );
+
+    expect(
+      const CopyWithValuesOptional(id: "test").copyWith(id: null).id,
+      null,
+    );
   });
 
   test('CopyWithProxy', () {